+++ classes/Coupon.class.php
@@ -6,6 +6,7 @@
 *   @copyright  Copyright (c) 2018 Lee Garner <lee@leegarner.com>
 *   @package    paypal
 *   @version    0.6.0
+*   @since      0.6.0
 *   @license    http://opensource.org/licenses/gpl-2.0.php
 *               GNU Public License v2 or later
 *   @filesource